feat: reformat json output to add options
BREAKING CHANGE: this changes the output of the json file to be keyed.
If you were using the metrics directly, all of the metrics will now be
under a key labeled "metrics". We also added an additional key called
"options" that gives you full access to all cli options that were passed
in.